feat: dedicated target per gateway specification with smaller-range groups
