+++ tooling/test/test.go
@@ -20,10 +20,10 @@ type SugarTest struct {
 
 type SugarTests []SugarTest
 
-func RunIfSpecsAreEnabled(
+func RunWithSpecs(
 	t *testing.T,
 	tests SugarTests,
-	required ...specs.Spec,
+	required ...specs.Leaf,
 ) {
 	t.Helper()
 
@@ -39,10 +39,10 @@ func RunIfSpecsAreEnabled(
 		return
 	}
 
-	Run(t, tests)
+	run(t, tests)
 }
 
-func Run(t *testing.T, tests SugarTests) {
+func run(t *testing.T, tests SugarTests) {
 	t.Helper()
 
 	for _, test := range tests {
